api/requirements.txt
requests==2.28.0
Django==4.0.5
djangorestframework==3.13.1
psycopg2==2.9.3